fix(learn): clear pending swipe when gesture commits; relocate SwipeDirection
