+++ frontend/src/components/learn/swipe-card-stack.tsx
@@ -2,11 +2,11 @@
 
 import type { RefObject } from "react";
 import { useCallback, useEffect, useImperativeHandle, useRef, useState } from "react";
-import type { SwipeDirection } from "@/app/learn/[cardgroupId]/learn-client";
 import { Button } from "@/components/ui/button";
 import { useReducedMotion } from "@/lib/use-reduced-motion";
 import { SwipeCard, type SwipeCardData } from "./swipe-card";
 import { SwipeProgressOverlay } from "./swipe-progress-overlay";
+import type { SwipeDirection } from "./types";
 
 /**
  * Delay between painting the swipe-progress overlay at full intensity and
@@ -18,9 +18,9 @@ const PROGRAMMATIC_COMMIT_DELAY_MS = 180;
 
 /**
  * Imperative handle exposed by SwipeCardStack so parents can trigger a swipe
- * programmatically (e.g. from rating buttons) without re-rendering the stack
- * through props. Keeping the handle minimal preserves React.memo bailouts on
- * the stack subtree across parent re-renders.
+ * programmatically (e.g. from rating buttons) without threading the
+ * direction through props. A stable handle + ref keeps the parent free of
+ * re-renders driven by per-frame drag-progress updates from useDrag.
  */
 export type SwipeCardStackHandle = {
   triggerSwipe: (direction: SwipeDirection) => void;
@@ -46,9 +46,10 @@ export function SwipeCardStack<TCard extends SwipeCardData>({
 }: Props<TCard>) {
   const activeCard = cards[0];
 
-  // Internal state for the in-flight gesture. Owned here (not lifted to the
-  // parent) so per-frame onSwipeProgress updates from useDrag do not cascade
-  // a setState through LearnClient's whole subtree on every pointer move.
+  // Drag-progress state ownership stays inside the stack, so the parent
+  // component never re-renders during a gesture — per-frame onSwipeProgress
+  // updates from useDrag would otherwise cascade a setState through
+  // LearnClient's whole subtree on every pointer move.
   const [swipeDirection, setSwipeDirection] = useState<SwipeDirection | null>(null);
   const [swipeProgress, setSwipeProgress] = useState(0);
   const reducedMotion = useReducedMotion();
@@ -88,6 +89,26 @@ export function SwipeCardStack<TCard extends SwipeCardData>({
     setSwipeProgress(progress);
   }, []);
 
+  // Gesture-completed swipes take precedence over any pending programmatic
+  // commit on the same card. Without this, a user who taps a rating button
+  // and then flicks the card before the PROGRAMMATIC_COMMIT_DELAY_MS timer
+  // fires would see onCardSwiped called twice for the same card — once via
+  // the gesture path, once via the late-firing timer — producing a duplicate
+  // mutation and an off-by-one completed counter. Clearing the timer here
+  // makes the gesture path absorb the queued programmatic commit.
+  const handleGestureCommit = useCallback(
+    (swipedCard: SwipeCardData, direction: SwipeDirection) => {
+      if (pendingCommitRef.current != null) {
+        window.clearTimeout(pendingCommitRef.current);
+        pendingCommitRef.current = null;
+      }
+      setSwipeDirection(null);
+      setSwipeProgress(0);
+      onCardSwipedRef.current(swipedCard as TCard, direction);
+    },
+    [],
+  );
+
   const triggerSwipe = useCallback((direction: SwipeDirection) => {
     const card = activeCardRef.current;
     if (!card) return;
@@ -111,8 +132,11 @@ export function SwipeCardStack<TCard extends SwipeCardData>({
 
     // Capture the card at trigger time so rapid clicks during the delay
     // window cannot route the commit at a card the user did not target.
+    // The active-card id guard is a belt-and-suspenders defence on top of
+    // handleGestureCommit clearing this timer when a gesture wins the race.
     pendingCommitRef.current = window.setTimeout(() => {
       pendingCommitRef.current = null;
+      if (activeCardRef.current?.id !== card.id) return;
       onCardSwipedRef.current(card, direction);
     }, PROGRAMMATIC_COMMIT_DELAY_MS);
   }, []);
@@ -185,7 +209,7 @@ export function SwipeCardStack<TCard extends SwipeCardData>({
           <SwipeCard
             card={card}
             isActive={index === 0}
-            onSwipe={(swipedCard, direction) => onCardSwiped(swipedCard as TCard, direction)}
+            onSwipe={handleGestureCommit}
             onSwipeProgress={handleSwipeProgress}
           />
         </div>
